Lsource eng lang attr (IS-130) and pos attr (IS-131) generation.
* In lsource elements, generate explicit "eng" language attribute
if the wasei flag is not true, even though "eng" is the default.
This seems match the current JMdict file XML.  Closes IS-130.

* In sense elements, suppress the generation of pos elements if
they exactly (in number, values, and order) match the pos elements
of the previous sense.  Again, this is for compatibilty with the
current JMdict file XML.  Closes IS-131.