jmxml bug causing jmparse.py to loose xrefs, fix jmextract.py.
[python/lib/jmxml.py]
If do_xref() was called multiple times (as it may be when
an xml entry sense had both and xref and an ant, the first
xrslv list attached to the sense will be overwritten (rather
that appended to) by the second list.  Fixed similar problem
in a number of other places (e.g. do_hist(), do_lsrc()) even
though in those place the function will be called only once
so problem wouldn't normally be seen.

In JmdictFile.read() strip uncode BOM if present on first
line, whether it is decoded unicode or utf-8.

[tools/jmextract.py, python/lib/jmxml.py]
Fixed a number of problems that prevented jmextract from
working.