Remove kwstatic.py, move Kwds class to kw.py.
Idea of python/lib/kwstatic.py was to provide a fast
way to initialize a Kwds instance when access to a
database not available.  However I concluded that extra
complexity, problems keeping it in sync with csv files,
make it better to just read the csv files directly when
needed.