entr.py: show entries in reverse order of most recent history.
When showing multiple entries, order entries within each group with
the same kanji/reading/sequence# by reverse order of date of most
recent history record.  This will usually place the most recently
edited version of an entry at top of the group of edits.  However,
if the first kanji or reading is changed in an edit, that version
of the entry will be placed according to its kanji/reading and not
with the other edits of the same seq number.