cgitbx.py: was failing to generate traceback
For a long time the cgitbx module was failing to generate a traceback
page as it was supposed to when running under Apache.  Instead, when
an exception occured in a cgi page, a Server Error page was generated
by Apache and the error log would contain nothing more the the un-
helpful message:
  AH02429: Response header name '<!--' contains invalid characters, aborting request
Problem was that the exception handler in cgitbx was executing but
printing a "reset" string before the "Content-Type: text/html" text.
Later versions of Apache are much stricter about valid headers and the
reset string was causing Apache to reject it.  Removing it restored
correct operation.