jmcgi.py: clear invalid sid values.
When an invalid sid value was received as a url parameter it would
not result in a valid session but it would not be cleared either,
resulting in its propagation in links and forms.  Although harmless
(since those forms would also recognise it as invalid) its presence
was confusing.  Clear it if it doesn't produce a valid session.

Also simplify assignment of 'sid_from_cookie'.