cgi/edsubmit.py: dont rollback submission when in dbg mode.
Formerly when the edsubmit.py cgi script was run with the url dbg=1
(debug) parameter, the submission would be rolled back rather than
committed to the database.  However this was quite confusing as the
submission acknowlagement page would not list any entry.  Since
there are options such as submitting to the test database or deleting
the entry afterwards, decided to complete the submission, even in
debug mode.