srchform.jinja: revise freq checkbox to set/clear the "P" values
When the P checkbox is cleared, all the "freq" checkboxes will
be cleared.  When the P checkbox is set, all the "freq" check-
boxes matching the set of P tags will be set, the others will
be cleared.

The previous version set and cleared only the checkboxes in the set
of P tags leaving the others unchanged but it was felt that checking
the P checkbox should provide the search criteria that matched the
meaning of P without being perturbed by other pre-existing settings.