tatoeba: map tatoeba en/jp ids to seq#s using diagonal mapping
Formerly one of the Tatoeba id numbers was multiplied by 1E7 and
the second added to it to create a JMdict seq#.  Problem with that
is the fixed multiplier value.  Scheme breaks if any Taoeba id
become larger than the multiplier.  When it was chosen 10+ years
ago the max Tatoeba id was ~1.5e6, it is now 6.5e6.
This commit switches to a diagonal mapping of the two Tatoeba id's
to a jmdict seq#.  This has no inherent upper bound.