Limit number of reverse xrefs to support Examples sentences.
This commit introduces known breakage (see below).

When the Tatoeba examples sentences are loaded in a JMdictDB database
and cross references to JMdict entries generated, some JMdict words
end up with a large number of reverse xrefs.  For example the word "の"
has close to 300K references to it from example sentences.

The result is excessive resource use (memory and time) when the
standard entr.py page attempts to format and display these reverse
xrefs in the course of displaying the page for "の".

This commit limits the total number of reverse xrefs to 100.  It is
not an acceptable fix because it does not distinguish between types
of reverse xrefs: some xrefs from other JMdict entries may not be
shown because they were "bumped" by the Examples sentences xrefs.
This will only be a problem in cases where the database has been
loaded with the Example sentences and xrefs to JMdict entries have
been generated.