xresolv.py: Create only one xref per xresolv item.
Previously, if an xresolv item did not specify any target sense number,
and the target entry had more than one sense, an xref to every sense was
generated.  We now generate an xref to only the first sense in this
scenario.  The rational is that only one of the multiple xrefs has a
reasonable chance of being correct, the others are almost certainly
wrong.  Either way, human intervention will be needed so it is better
to provide only a single xref rather than cluttering the target entry
with many identical reverse xrefs.

Temporarily use xrefs.notes to record example sentences sentence-text,
priority, and no-explicit sense values.  These will get their own
fields in the xref table but for now this hack is an interim measure.