xresolv.py: unsuccessful attempt to speed up
Earlier version of this program ran in less then 10m to resolve the
Examples file xrefs against the jmdict corpus.  Version previous to
this one runs in ~50m.  This version was an attempt to speed up by
getting the xref target candidates and all the info needed to choose
the best candidate and create the new xrefs with a single query rather
than reading and creating new Entr objects, one by one, for that purpose.
however this version still takes ~45m.

Problem seems to be with the INSERT statement used to create the resolved
xrefs; in particular with the ON CONFLICT clause.  When it is removed and
with no other changes execution time drops back down to ~10m.

Because this version was experimental, many of the command line arguments
are non-functional.