xresolv.py: simplify the queue-to-file adapter class
Replace the IterableQueue and IteratorFile classes (both scavenged
from the internet) with our own, simpler FileQueue class.  This seems
to fix the terminating problem, at least when there are no fatal
errors in the writer threadds.
Not much difference (~10-20s) when running with 5 threads vs 1 so
not sure how much the threading buys.
Expanded comments.