Remove version number derived from Mercurial revision number.
We were setting __version__ in most modules and command line tools
from the Mercurial revision number and date.  These were extracted
via Mercurial's keyword expansion feature.  This was a holdover from
when CVS was used and in 15 years I've never once made use of this
info.  Since porting it to Git is burdensome, will drop it.